requirements.txt
certifi==2021.5.30
chardet==4.0.0
future==0.18.2
idna==2.10
line-bot-sdk==1.19.0
requests==2.25.1
urllib3==1.26.5